# MAINT: Speed up numpy.nonzero.