… numpy/core/src/multiarray/lowlevel_strided_loops.c.src
… numpy/core/tests/test_numeric.py