This PR solves issue [11569](https://github.com/numpy/numpy/issues/11569). @tylerjereddy also requested optimizations for large boolean arrays. I have added optimizations to numpy.nonzero for contiguous and aligned numpy arrays. Generally, the speedup is at least 2 times and the maximum speed gain observed is about 8 times. Currently, the optimizations are exclusive to all int types, float types and bool type for arrays of dimension 1 to 3. If needed, optimizations for higher dimensional arrays can be added. Test cases have been added.

The following code illustrates some of the speed gains:
```
import numpy as np

x = np.random.rand(1000000) > 0.5
%timeit np.nonzero(x)
666 µs ± 2.04 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
4.33 ms ± 17.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.rand(1000,1000) > 0.5
%timeit np.nonzero(x)
1.18 ms ± 2.68 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
9.55 ms ± 32.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.rand(100,100,100) > 0.5
%timeit np.nonzero(x)
1.48 ms ± 3.55 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
9.57 ms ± 44.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.rand(1000000) > 0.9
%timeit np.nonzero(x)
655 µs ± 2.46 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
1.77 ms ± 7.87 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [Without optimization]

x = np.random.rand(1000,1000) > 0.9
%timeit np.nonzero(x)
1.02 ms ± 2.46 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
4.37 ms ± 12.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.rand(100,100,100) > 0.9
%timeit np.nonzero(x)
1.22 ms ± 9.38 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
3.9 ms ± 65.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.randint(0,9, size=(1000,1000), dtype=np.int32)
%timeit np.nonzero(x)
1.69 ms ± 2.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
7.83 ms ± 17.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.randint(0,2, size=(1000,1000), dtype=np.int32)
%timeit np.nonzero(x)
1.49 ms ± 5.86 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
10.4 ms ± 31.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.randint(0,9, size=(1000,1000), dtype=np.int64)
%timeit np.nonzero(x)
2.3 ms ± 6.37 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
8.15 ms ± 29.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]

x = np.random.randint(0,2, size=(1000,1000), dtype=np.int64)
%timeit np.nonzero(x)
1.93 ms ± 4.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) [With optimization]
10.7 ms ± 22.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) [Without optimization]
```