@charris 2021-02-08
Please note the NumPy [`C_STYLE_GUIDE`](https://numpy.org/neps/nep-0045-c_style_guide.html).