@touqir14 2021-02-09
@seberg ,I have pushed recent commits that optimize for arrays and views that are F-contiguous, including those arrays with a fortran based memory layout. So, currently, any aligned and contiguous array will enjoy the optimization. Thats it from me for now. You can continue with the review. 

Regarding the speed on F-contiguous arrays, the speed matches that of C-contiguous arrays given the shapes are transposed. So, a C-contiguous array with shape: [m, n] will attain the same speed as an F-contiguous array with shape: [n, m] given they have the same values.