@seberg 2021-02-12
Ok, I think the speed up is probably worth the additional churn.  

I was just now surprised by how much slower nonzero is compared to the comparison itself, i.e. `trues = x > 0.5; np.nonzero(trues)` compared to `np.nonzero(x > 0.5)`, that realigns my view on this entirely. I think we definitely should do it.
```
In [8]: x = np.random.rand(1000,1000)
In [9]: %timeit x > 0.5
487 µs ± 59.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
In [10]: %timeit np.nonzero(x > 0.5)
9.24 ms ± 103 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

There are a few things left to do, but I could help a bit if you are running low on enthusiasm:

1. We should have benchmarks for all the cases you explicitly made faster. And I would like to include at least one or two also for `np.nonzero(double_array > 0.5)` style call, because I think that is effectively the more common case. I assume this means we need to add new ones to `numpy/benchmark/`.
2. You have pretty big macros right now. Those macros should be functions written in the NumPy "templating" language.  The easiest way here will be to move those functions into `lowlevel_strided_loops.c.src`.  That should will likely allow you to combine `F` and `C` order functions into a single one. (Not sure about functions for the number of dimensions.)

Question to be sure (maybe it is obvious from the tests, in which case, don't bother):
* Optimizing `F` order as well is nice, but I think `np.nonzero` might always return C-order indices right now. We must not change that.  We should have a test for it, and if the current code should change it, maybe just pull it out again for the moment.

Backburner:
* There are still some C-style nits (long lines, missing brackets)
* I am a bit curious (if I am not off track again), whether skipping the `count_nonzeros` and instead using `realloc` is useful. I guess it should not be, since that is so much faster than the actual `nonzero` core?
* The huge speedup seems almost "too large", I am wondering if the generic path is not implemented much slower than it should be.