@touqir14 2021-02-12
I am actually not acquainted with the numpy template language. It is possible to combine the f and c style loops using more macros but I intentionally left them separated for better code readability and maintainability. Still if we must combine the f and c style loops together, how about we try to do that in a fresh PR? We can then also think about ways to generalize to higher dimensional loops? 

The output of ```np.nonzero``` is a tuple of 1D arrays. F style or C style does not affect the output. Is that what you were concerned about?