@touqir14 2021-02-12
The use of ```count_nonzero``` actually benefits the nonzero function's speed for sparse cases. The ```count_nonzero``` probably accounts for 1/6th or less time which I think is worth being included for the sparse cases. For higher the dimensions the lower the overhead is due to ```count_nonzero```. The SIMD optimization from [18183](https://github.com/numpy/numpy/pull/18183) has really been useful to keep the overhead low.