@seberg 2021-02-12
The `c.src` are a bit unique, but not very tricky see here for details: https://numpy.org/doc/stable/reference/distutils_guide.html#other-files.  I am not sure if it makes sense to combine fortran/C into a single chunk, so fine with leaving it for now.

My question about the order is the output is about this example:
```
In [4]: np.nonzero(np.ones((2, 2)))
Out[4]: (array([0, 0, 1, 1]), array([0, 1, 0, 1]))

In [5]: np.nonzero(np.ones((2, 2), order="F"))
Out[5]: (array([0, 0, 1, 1]), array([0, 1, 0, 1]))
```
I.e. its easy to imagine that the last might instead return:
```
(array([0, 1, 0, 1]), array([0, 0, 1, 1]))
```
which is also correct but different.