@touqir14 2021-02-12
Actually, this implementation returns the second one : ```(array([0, 1, 0, 1]), array([0, 0, 1, 1]))``` . This is unavoidable if we do not want to lose performance. For f-style arrays, f-style loops are always faster. Is it really important to preserve the order? In fact, if those indices are somehow later used to index that f-style array for some BLAS operation, this should be faster than having the output given in a c-order.