@seberg 2021-02-12
I have to think about it, but it is a subtle change that I would prefer we don't include here yet (to ensure that the rest doesn't get stalled because of that!).
The solution might also be to add `np.nonzero(..., order="C")`. I.e. an `order` parameter  (we could limit support to only `C` or `K` or so, where `K` would mean "I don't care").