@seberg 2021-02-12
I meant progress on this PR getting stalled, because by the question of changing the output order being OK.  A new `order` keyword argument is likely pretty uncontroversial, but also feels like it may be a good followup.