@touqir14 2021-02-12
@seberg , I am getting this error after modifying the low level strided loops file : ```ValueError: In "numpy/core/src/multiarray/lowlevel_strided_loops.c.src" loop at line 1: no definition of key "elsize" ``` and I haven't modified anything. Its probably something common and there is a fix.