@seberg 2021-02-12
@touqir14 sounds like the "preprocessor" is breaking on something, maybe you are using @elsize@, but did not properly define a loop `elsize=1,2,4` (or because you edited some `/**begin repeat` that is not quite right and a later one is confused by it. If you can show the diff, I can take a look.